assignment3/fastapi-auth-app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymysql
python-jose
passlib[bcrypt]
python-dotenv